pyproject.toml
[project]
name = "pubmed-fetcher"
version = "0.1.0"
description = ""
authors = [
    {name = "Fiona Lazarus"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython (>=1.85,<2.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
flake8 = "^7.1.2"

[tool.poetry.scripts]
get-papers-list = 'pubmed_fetcher.main:main'

